common-lib: Rename Serial to Uart
[Major]
Rename Serial to Uart
- When expending serial feature, name can be confused.

Refactoring Uart
- Hide implementation -> Make clean header
- Improve code quality

[Minor]
Fix thread unit test failure on Linux

task: COMMO-34
