--- include/common/communication/driver/I2C.hpp
@@ -1,49 +1,0 @@
-/**********************************************************************
-MIT License
-
-Copyright (c) 2025 Park Younghwan
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-**********************************************************************/
-
-#pragma once
-
-#include "CommonHeader.hpp"
-#include "common/communication/driver/BaseDriver.hpp"
-
-namespace common::communication
-{
-struct I2CInfo : public DeviceInfo
-{
-    enum Speed : uint32_t
-    {
-        Low         = 0x00, // 20KHz
-        Standard    = 0x01, // 100KHz (default)
-        Fast        = 0x02, // 400KHz
-        High        = 0x03, // 750KHz
-    };
-
-    I2CInfo() 
-        : DeviceInfo(DeviceInfo::I2C) {}
-
-    Speed _speed = Standard;
-    uint8_t _devAddr = 0x00;
-    uint8_t _regAddr = 0x00;
-};
-} // namespace common::communication
